requirements.txt
numpy
pandas
scikit-learn
matplotlib
pyyaml
jupyter
nbconvert
ipykernel
